refactor styles
