+++ src/DataTable/ui/Sort.module.scss
@@ -1,0 +1,43 @@
+.sortingContainer {
+  position: relative;
+  display: flex;
+  margin-left: 5px;
+}
+
+.sortingButton {
+  display: flex;
+  position: absolute;
+  top: -1rem;
+  left: 0;
+  margin: 0;
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0;
+  color: var(--sui-datatable-white);
+}
+
+.sortDropdown {
+  position: absolute;
+  top: -1.8rem;
+  left: 2rem;
+  background: var(--sui-datatable-black);
+  border: 1px solid var(--sui-datatable-border);
+  border-radius: var(--sui-datatable-border-radius-small);
+  z-index: 10;
+  display: flex;
+  flex-direction: column;
+}
+
+.sortOptionButton {
+  border: none;
+  background: none;
+  padding: 5px;
+  text-align: left;
+  cursor: pointer;
+  width: 100%;
+}
+
+.sortOptionButton:hover {
+  background-color: var(--sui-gray-150);
+}
